+++ docs/conf.py
@@ -121,13 +121,13 @@ html_static_path = ["_static"]
 
 # The name of an image file (relative to this directory) to place at the top
 # of the sidebar.
-html_logo = ""
+html_logo = "_static/img/SaltProject_altlogo_teal.png"
 
 # The name of an image file (within the static path) to use as favicon of the
 # docs.  This file should be a Windows icon file (.ico) being 16x16 or 32x32
 # pixels large. Favicons can be up to at least 228x228. PNG
 # format is supported as well, not just .ico'
-html_favicon = ""
+html_favicon = "_static/img/SaltProject_Logomark_teal.png"
 
 # Sphinx Napoleon Config
 napoleon_google_docstring = True